Created skeleton code for log in and register
