+++ imdb/UI/MVP/View/LoginViewController.swift
@@ -1,0 +1,13 @@
+//
+//  LoginViewController.swift
+//  imdb
+//
+//  Created by User01 on 23/7/23.
+//
+
+import Foundation
+import UIKit
+
+class LoginViewController: AuthenticationViewController {
+  
+}
