+++ imdb/UI/MVP/View/RegisterViewController.swift
@@ -1,0 +1,13 @@
+//
+//  RegisterViewController.swift
+//  imdb
+//
+//  Created by User01 on 23/7/23.
+//
+
+import Foundation
+import UIKit
+
+class RegisterViewController: AuthenticationViewController {
+    
+}
